+++ integration_tests/async_await/src/main.rs
@@ -1,4 +1,4 @@
-use juniper::{graphql_value, RootNode, Value};
+use juniper::{graphql_value, GraphQLError, RootNode, Value};
 
 #[derive(juniper::GraphQLEnum)]
 enum UserKind {
@@ -34,8 +34,7 @@ impl User {
     }
 
     async fn delayed() -> bool {
-        let when = tokio::clock::now() + std::time::Duration::from_millis(100);
-        tokio::timer::delay(when).await;
+        tokio::time::delay_for(std::time::Duration::from_millis(100)).await;
         true
     }
 }
@@ -61,8 +60,7 @@ impl Query {
     }
 
     async fn delayed() -> bool {
-        let when = tokio::clock::now() + std::time::Duration::from_millis(100);
-        tokio::timer::delay(when).await;
+        tokio::time::delay_for(std::time::Duration::from_millis(100)).await;
         true
     }
 }
@@ -72,14 +70,8 @@ struct Mutation;
 #[juniper::graphql_object]
 impl Mutation {}
 
-fn run<O>(f: impl std::future::Future<Output = O>) -> O {
-    tokio::runtime::current_thread::Runtime::new()
-        .unwrap()
-        .block_on(f)
-}
-
-#[test]
-fn async_simple() {
+#[tokio::test]
+async fn async_simple() {
     let schema = RootNode::new(Query, Mutation);
     let doc = r#"
         query { 
@@ -95,9 +87,9 @@ fn async_simple() {
     "#;
 
     let vars = Default::default();
-    let f = juniper::execute_async(doc, None, &schema, &vars, &());
-
-    let (res, errs) = run(f).unwrap();
+    let (res, errs) = juniper::execute_async(doc, None, &schema, &vars, &())
+        .await
+        .unwrap();
 
     assert!(errs.is_empty());
 
@@ -118,6 +110,35 @@ fn async_simple() {
             },
         }),
     );
+}
+
+#[tokio::test]
+async fn async_field_validation_error() {
+    let schema = RootNode::new(Query, Mutation);
+    let doc = r#"
+        query {
+            nonExistentField
+            fieldSync
+            fieldAsyncPlain
+            delayed
+            user(id: "user1") {
+                kind
+                name
+                delayed
+            }
+        }
+    "#;
+
+    let vars = Default::default();
+    let result = juniper::execute_async(doc, None, &schema, &vars, &()).await;
+    assert!(result.is_err());
+
+    let error = result.err().unwrap();
+    let is_validation_error = match error {
+        GraphQLError::ValidationError(_) => true,
+        _ => false,
+    };
+    assert!(is_validation_error);
 }
 
 fn main() {}