+++ juniper_rocket/CHANGELOG.md
@@ -1,6 +1,11 @@
 # master
 
 - Compatibility with the latest `juniper`.
+- Rocket integration does not require default features.
+
+# [[0.5.2] 2019-12-16](https://github.com/graphql-rust/juniper/releases/tag/juniper_rocket-0.5.2)
+
+- Compatibility with the latest `juniper`.
 
 # [[0.5.1] 2019-10-24](https://github.com/graphql-rust/juniper/releases/tag/juniper_rocket-0.5.1)